+++ app/src/main/java/com/sopt/now/data/repoImpl/UserRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.sopt.now.data.repoImpl
 
-import com.sopt.now.data.repository.UserRepository
+import com.sopt.now.domain.repository.UserRepository
 import com.sopt.now.data.service.UserService
-import com.sopt.now.presentation.user.UserInfo
+import com.sopt.now.domain.entity.UserInfo
 
 class UserRepositoryImpl(private val userService: UserService) : UserRepository {
