+++ app/src/main/java/com/sopt/now/presentation/login/LoginActivity.kt
@@ -14,7 +14,7 @@ import com.sopt.now.presentation.signup.SignUpActivity
 import com.sopt.now.databinding.ActivityLoginBinding
 import com.sopt.now.data.dto.request.RequestLoginDto
 import com.sopt.now.data.repoImpl.AuthRepositoryImpl
-import com.sopt.now.data.repository.AuthRepository
+import com.sopt.now.domain.repository.AuthRepository
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
