+++ app/src/main/java/com/sopt/now/presentation/login/LoginViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.sopt.now.core.view.UiState
 import com.sopt.now.data.dto.request.RequestLoginDto
-import com.sopt.now.data.repository.AuthRepository
+import com.sopt.now.domain.repository.AuthRepository
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
